add decrypt command to decrypt individual share file
The decrypt command will decrypt a share file produced using the --gpg-keys
split option.

The command will read the share file, decrypt the payload (provided you have
access to the proper gpg key), and produce a share file with the decrypted
payload. This share file can be used in a future restore operation.